Load subcategories + appearances for ViewState.
TFS#911403: FeatureSymbologyOverrides need to know the set of
subcategories that are visible in the view given the set of categories
in the category selector. Additionally, need the base subcategory
appearances for each.
NEEDSWORK: Currently ignoring overrides on DisplayStyle, and not
updating the subcategory info when an initially-disabled category
becomes enabled.