Merged PR 6409: Enable imodeljs-clients integration tests to work with iModelBank
# Integration tests
The **clients** tests no longer assume that they are working directly with Connect. Instead, they work with a "project abstraction" to do things like authenticating a user or discovering and working with projects.
Likewise, the tests no longer assume that the imodel server is a hub that serves multiple iModels. Instead, they work with the project abstraction when discovering imodels.
The tests no longer assume that there is a single IModelClient for any or all iModels. Instead, they ask for the client to use for each iModel.

I have made some progress  toward changing the imodeljs-backend integration tests to be able to work with iModelBank. **TBD**: each test suite must *create* the iModels that it wants to work with, rather than assuming that they are already there in the server. (The clients tests create their own iModels. The backend mock assets contain all of the data needed to do this for the backend tests.)

# IModelBankFileSystem
Started defining how an iModel and its history are stored on disk, for use by iModelBank. Wrote an implementation of the project abstraction to set up and work with iModel directories for use by iModelBank. These can be used by real servers.

Wrote wrappers to run iModelBank servers for testing purposes.