Fixed the HTTP Route when opening ReadWrite IModelConnection. The path doesn't include the changeSetId field anymore, and the appMode is correctly set to to "2" (WorkGroupEdit) instead of "1" (MilestoneReview)
The path before the fixes:
`${prefix}/${appTitle}/${appVersion}/mode/${appMode}/context/${contextId}/imodel/${iModelId}changeset/${changeSetId}/${operationId}`,
where appMode = "1" (MileStoneReview)
The path after the fixes:
`${prefix}/${appTitle}/${appVersion}/mode/${appMode}/context/${contextId}/imodel/${iModelId}changeset/${changeSetId}/${operationId}`,
where appMode = "2" (WorkGroupEdit)

The IModelToken now includes the OpenMode field to convey the appMode to the RPC code.

The changeSetId field is omitted in the route in the case of ReadWrite connections since
the connection is generally expected to be at the latest version and not some specific changeSet.
Also, for the first version (before any changeSets), the changeSetId in the route is arbitrarily set to "0"
instead of an empty string, since the latter is more un-intuitive for a route.

However, in all other use cases, including the changeSetId held by the IModelToken itself, the changeSetId of ""
(i.e., empty string) signifies the first version - this is more intuitive and retains compatibility with the majority
of use cases.