Merged PR 8405: Added tests for index imports
#37538 Test added, exports are imported from Index and tested against explicitly imported modules to ensure equality.

"Export from" was changed to "Export as" because the only exported member of the Constants class was named "default". This caused those imports to be imported as "default", which has the chance to be redefined by other imports. They are now exported as ECStringConstants, the name of the default export class.

Related work items: #37538