Merged PR 11546: schema rules/validation
First version for rules/validation. One key thing missing is the ability to chain/link diagnostics together. In the second version (in the other branch), the design sort of eliminated the ability to do this which is why I think this design may be better, but I like some of the things in the second branch better....