Merged PR 16541: Fix for 74344 (Tiles not loaded when root document is not at root of blob.
The previous code assumed that the root document was always located at the root of the blob. Although this is true of Reality Data uploaded using DgnV8Converter it is false for most applicaitons and specifically for Context Capture products that place the Reality data inside the 'Scene' folder. The code change obtains the root document name and retains the path part of this root document to be appended to all tile requests (that are identified relative to root document)

Added quite a few comments to clarify operation and identified a few TODO things.

Related work items: #74344