Merged PR 18051: Implemented a Web Workder framework.
Implemented a Web Worker framework for use in frontend applications. The ostensible use case was decoding jpeg data in GltfTileIO, but the performance improvement was judged not worth the extra complexity and complication of getting it working on all browsers, so it is commented off there. But WebWorkerManager was tested successfully and together with webworker/WebWorker.ts illustrates how to queue requests to a pool of worker threads and service requests in the receiving web workder.
Also made some documentation corrections.