Merged PR 19514: Planar Classification Support
Planar classification classifies using a texture that is projected along plane normal (currently the X-Y plane) -- This is more efficient than the volumetric classification as it does not require complex, closed volume tiles and can do classification with standard tiles.  There is no need to separate primitives or make a single tile.