Merged PR 19480: Change path to config-loader module
In applications that use rush/pnpm outside of the imodeljs repo, the existing path would not exist (it is not symlinked).

This change requires @config-loader as a dependency in the cwd as well, so I'm not sure if its the best approach. However, there is precedent in this file already with @bentley/imodeljs-frontend and the assumption that the current path exists is also not optimal.