Merged PR 21521: Fix tile LOD issue
If a tile had a child tile whose content had been loaded and deemed to be empty, selectTiles() would incorrectly decide to draw the parent in place of its children. This produced missing tiles if the parent was a root tile with no graphics, or lower-resolution graphics if the parent had graphics.
This was a regression caused by no longer making empty tiles with children have a maximumSize of 0 (which made them "undisplayable", which made us immediately try to load their children's content).

Related work items: #97270