Merged PR 23872: Support per-model category visibility (for visibility tree)
Viewport.perModelCategoryVisibility allows, for a set of models, the visibility of a set of categories to be overridden in the context of a viewport. Enabling a category for a model enables ALL subcategories of that category within that model.
Adds a new Viewport.onViewedCategoriesPerModelChanged event.

The already-complicated logic for asynchronously loading subcategory info for viewed categories has become more complicated in order to enforce order of request handling.

Added lots of Id64-related APIs.

Related work items: #100283