Merged PR 25069: Fix default hidden edge pattern
NOTE: The converter leaves "pattern" undefined in the DisplayStyle json if hidden edges are supposed to use the default dashed line pattern. If MSTN display style does NOT override the hidden edge pattern, the converter overrides it to be SOLID. This is incorrect but no mechanism currently exists in the converter or the JSON representation to specify "use element pattern for hidden edges". The incorrect default pattern is a bigger problem so choosing to err on that side instead.