Merged PR 28444: Cleanup optional and token-derived ULAS REST API parameters
The ULAS REST API that iModel.js uses to log usage allows for many properties to left empty. In some cases, these empty properties will be given values derived from the access token being used to log the usage. We defer the setting of these properties to the ULAS server.