Merged PR 29041: Improve ULAS error logs
The current ULAS error logs merely inform of an error & supply the imodelDb token, which may not even be the cause of the error.
More information, such as the Error code and message, will give context to the errors.

Additionally, the projectId/contextId was not being properly conveyed to ULAS in IModelDb.open() calls. This has been remedied and clarified by adding contextId as a constructor parameter to all LogEntry classes.