Merged PR 31617: Make IModelDb.insertElement never return Id64.invalid, throws instead
This method used to both throw and potentially return invalid id. That makes no sense. It now only throws, so people don't have to check the return value.

I also made several static methods on Element and Model @beta that were previously (accidentally) @public. That's a breaking change, but I know for a fact that nobody could be impacted by this, since they are only for domain implementers.

@<Affan Khan> I removed all the tests for "isOpen" in IModelDb. Why would they be there? How did I not see the PR where they were added?