Merged PR 33394: Fix category picker; reinstate isOpen check in pollTileContent.
After asynchronously loading subcategories we must mark the category selector as having changed so that we will recompute the feature symbology overrides. We previously only did this if caller wanted all subcategories enabled, which display-test-app does not.
Keith removed a check on IModelDb.isOpen in pollTileContent. This is important because pollTileContent is called in a setTimeout callback and the iModel can be closed before that gets invoked.