Merged PR 33407: Fix model selector view sync problem
When the display of a model is changed through the model selector, the view doesn't get updated automtically untill the user interacts with the view.

Please note that most of the changes below are just formatting changes (I have a prettier plugin installed in VSCode). The actual change is in the _onModelsChecked starting on line 204. Here I use the more comprehensive `Viewport.addViewedModels` method that takes care of loading any unloaded models and also makes an `invalidateScene()` call. For turning a model of I call `ViewPort.changeModelDisplay` directly that invalidates the scene if requried.

Related work items: #137708