Merged PR 34788: Add exportPartGraphics and export-gltf
Add the capability for exportGraphics to discover references to GeometryParts and pass those references back to clients. Also add exportPartGraphics to allow clients to create meshes for GeometryParts selectively.

Also add a new test app, export-gltf. This creates GLTF 2.0 files that leverage the new instancing API and serves as both documentation and a "dog fooding" implementation.