Merged PR 34295: Flaky test failing on linux
This is a guess fix where i assume that it is happening due to number VCPU  on linux CI vm. I have reduce the number thread used by concurrent to a fix of 2 thread. Instead of initializing it to the number of available vCPUs which seem to 10 per VM in linux CI vm.