Merged PR 36611: Fix viewport image blending
If we need to resize the image returned from Viewport.readImage, we fill a resized canvas with the view's background color and then draw the image onto it. Any translucent pixels would blend with that background color, which is incorrect. Added 'preserveAlpha' option to 3 related functions; if false, alpha channel is set to 255.
Left the default behavior if argument not supplied (ImageUtils.ts contains general-purpose functions; the blending issue is specific to the scenario of pre-blended viewport images).