Merged PR 33450: First draft of Capabilities.md
The intention here is to produce an overview of all capabilities of iModel.js in the documentation.  The list is grouped by category and within each category a list of capabilities.  For each capability we should have a short name and 2-3 sentence description.  As much as possible, the description should link to learning or reference documentation for more details.

I will add comments to the pull request assigning sections to people.  Please add/remove capabilities and add descriptions.

@<Lu Han>  and I filled in 'Visibility Control' as an example.