Merged PR 29375: Removed seq port config option from FluentdConfig.
`seqServerPort` property was useless and hence misleading because it was not taking any effect. Rather if a port is needed to be specified, it can be provided with `seqServerUrl` like for example `localhost:5341`.

For more details on this feature, please view the initial PR for this feature [here](https://bentleycs.visualstudio.com/iModelTechnologies/_git/imodeljs/pullrequest/10458?_a=overview).