Merged PR 41138: Support animation of classifiers.
Animation and classification were previously (unnecessarily) mutually exclusive.  The Microsoft  campus demonstration requires the resymbolization of the building footprints according to schedules based on historical data.   This change does not include the tile generation so the animation tiles are just generated as primary for now (good enough for demo) -- When the native code is updated to support animation and classification the animation ID can be combined with classifier more elegantly.