Merged PR 42944: Add workaround to support reality model tiles at large distances from origin without accuracy loss.
If a b3dm tile is found that is far from the origin then the tile center is treated as the RTC and the tile center is subtracted from vertices when read -- or in this case to the quantization origin.   This removes the jittering as the 7 digit plus coordinate values were not accurately represented by 32 bit floats.