Merged PR 44286: Check for redundant lock/code requests, fix tests
Now that iModel.js no longer allows pushing of changes without first acquiring locks and codes, we have caught a few tests that were incorrect.

Also, as a short-term strategy, we also detect and do not choke on redundant requests for locks and codes. This could arise in an existing app if the app uses the low-level iModelHub API to acquire locks and codes and then adds the now-required call to ConcurrencyManager.request. The request method is driven by the locks and codes disclosed by the element handlers (in the addon). The request method must detect which resources are already held and must not trigger an error by making a redundant request for them.