Merged PR 45530: Replace in-line script tag in index.html's with referenced checkbrowser.js
The script tag in index.html needed to check the browser for compatibility was too verbose to inflict on everyone, so I made a new checkbrowser.js that does approximatlely the same thing. This way it can be improved if necessary and everyone gets the advantage of any changes.
Refinement in Tool class to always set i18n and eliminate runtime check. Also added i18n argument to registerModule.