Merged PR 49990: Simplify and refine shadow map
Removed unnecessary logic for detecting when the shadow map needs to be regenerated.
Do not (attempt to) wait for the *best* set of tiles for the current shadow frustum - just use the best *available* - when the *best* ones arrive, we will update the shadow map accordingly.
When switching from 3d view to 2d view, free the shadow map resources.

Related work items: #159448, #203120