Merged PR 51127: Address deprecation warnings in GeometryStream.test.ts
I was taking a quick pass to try to knock off some deprecation warnings in our code. I ran into a bunch related to `GeometryStreamIteratorEntry`

The `@deprecated` tag typically means that the API item will be removed in a future release. After talking to Brien, the determination was that the properties marked as `@deprecated` in the `GeometryStreamIteratorEntry` class should not be removed so I deleted the `@deprecated` tag instead.