Merged PR 53252: Clip preloaded low resolution tiles to the higher resolution range of the children they are replacing.
If a low resolution tile is displayed because a child meets SSE but is not loaded then clip the range of that tile to display.  This avoids having low resolution terrain temporarily overwrite higher resolution tiles.   This greatly improves the stability of the terrain display and the transparent display as well.