Merged PR 53642: Initializing OidcBrowserClient before setting it in UiFramework
Problem:
On application start, we are getting BentleyError: Not signed in (will attach screenshot), but we still are getting into application (we are using refresh_token). After looking around iModelJs code, I noticed that it is trying to get access token from oidc client before initializing it, so it will throw everytime.

Solution:
Initializing oidc client before setting it in UiFramework, which is trying to get access token. In this case it will only thrown if user is actually not signed in.

![Screenshot_1.png](https://dev.azure.com/bentleycs/2c48216e-e72f-48b4-a4eb-40ff1c04e8e4/_apis/git/repositories/692a403b-7ecc-4797-adb1-e6f5c7080aa9/pullRequests/53642/attachments/Screenshot_1.png)