Merged PR 53835: Basic hypermodeling display
Implements the ability to display a 2d model in a spatial view oriented with the section plane.
Currently displays the *drawing* view. We're actually supposed to display the *sheet*, with the drawing view attachment and annotations. For that we need the converter to preserve more information - in particular, the clipping boundary on the sheet (this differs from the drawing boundary; it is expanded to include sheet annotations).

Related work items: #213431