Merged PR 55098: UI: Extend MutableTreeModel API to support node inserting, removing
Added additional methods for changing TreeModel:
 * removeChild(parentId, childId) - removes specified node
 * insertChild(parentId, childInput, offset) - inserts node at specified position without overriding existing nodes
 * getChildOffset(parentId, childId) - returns child offset in children array.

Changed setChildren method to create children array for parent if it does not exists.

Fixed problem that children where added to model for collapsed parent when disposeOnCollapse is enabled. This would happen if there are pending load request and parent node is collapsed. Existing children were removed but children from pending requests were added to model. This fix breaks useControlledTreeUnifiedSelection hook, but that will be fixed in a separate PR.

Related work items: #215739