Merged PR 64880: Add PlanProjectionSettings.enforceDisplayPriority
This flag indicates for a given model whether or not to use subcategory priority to control display order.
If set, we will generate different tiles that group the graphics according to subcategory.
In preparation for adding new types for working with layer priority, break up monolithic System.ts into sensible smaller files.