Merged PR 66413: IModel frame lifecycle events
Added frame lifecycle callbacks needed by 3dmaps plugin to be able to hook into the imodeljs rendering pipeline.
Also added id to the canvas elements because 3dmaps need to be able to identify the drawing canvas.

I marked the exported interfaces as "internal" not sure if this is the correct approach.
Please review.