Merged PR 70791: Cherry-pick: Merged PR 70528: Change scss imports to use a `~` to support Yarn workspace
Change scss imports to use a `~` to support Yarn workspace hoisting

The current way of handling scss imports breaks when the node_modules is not in the expected format.  This goes against the module resolution strategy used by sass loader and the way Yarn workspaces, and other mono-repo solutions, work.