Merged PR 73188: Fix Support for dialogs with no header and custom move logic
For the dialog component, support dialogs that disable the default header and want to implement custom move logic by updating the x and y props. This was fixed by only handling the pointer move events when a resize or move operation starts instead of on mount. Also now we only update the relevant state in the pointer move event handler based on if we are moving or resizing.

Before this change, the _handlePointerMove handler is always called and it always set the state

`positionSet: true `

that caused our custom dialog to no longer be able to change the dialog position by updating the x and y props.

Related work items: #266203