Merged PR 75005: Update Extension loading to support the Extension Service
Hey everyone, this is the first of a few changes to support adding Extensions via the Extension Service.  I'll be making further updates in addition to this PR to support the new build system for iModel.js and the Extensions.  Then we'll also be cleaning up the docs in these few files.