Merged PR 80395: Remove named clip code, we ended up not wanting to save clips independent of saved views.
Remove named clip code, we ended up not wanting to save clips independent of saved views.

After discussing with Carl and Joe Z it was agreed that it would be a good idea to remove this code that we didn't end up using once the named sections widget was removed from DR (and it has been).