Merged PR 82587: Always set planar flag for 2d geometry.
Element contains a region with no outline; plus a bunch of line segments that are supposed to stand in for the region outline. The segments z-fight with the surface because they are not real edges and they are not recognized as planar, whereas the surface is. Everything in a 2d model is planar so make sure to always set that flag.

Related work items: #303334