Fail on expired sas url
As of now if sas url is expired we download the 403 extend response which is xml and save as changeset file or checkpoint file. It would fail later on but user does not get instant error and thus make debugging harder.

This happen frequently if you attempt to download thousand of changeset by fetching changeset definition first and since download url expire in 5 minute at some point in downloading these changeset 10 at a time would start failing but as of now it does not fail but save xml as target file.