Merged PR 83325: Restore old surface discard logic.
This PR: https://dev.azure.com/bentleycs/iModelTechnologies/_git/imodeljs/pullrequest/80497?_a=overview
Attempted to narrow the scope in which we would discard surfaces in favor of the contents of the pick buffer, because of a specific view Shaun had in which zooming out far enough caused the edges of "roadway layers" just below the road surface to show through.
It turns out however that the benefits of the old logic outweigh the risk of such anomalies: with those changes, edges of adjoining walls z-fight or disappear, as do area patterns.
Reverted those changes; we will need to investigate alternative options for addressing Shaun's anomaly (e.g., computing a depth tolerance based on the frustum).