Merged PR 83562: Delay when static resources are unpacked using webpack
An issue arose with a project using the "clean-webpack-plugin" together with CopyBentleyStaticResourcesPlugin. The issue was that CopyBentleyStaticResourcesPlugin unpacks the static resources to build directory before "clean-webpack-plugin" cleans the output, thus the static resources are removed.

Fixed by executing static resource unpacking in "compiler.hooks.emit" instead of "compiler.hooks.beforeRun".