Merged PR 84110: Make Entity.forEachProperty public and change the default parameters to better match typical use.
`Entity.forEachProperty` has moved from beta to public. To better match typical use, the default value of the `includeCustom` parameter has changed. This makes it so that most consumers don't need to understand the distinction between *custom* and *auto-handled* properties.

The internal code that requires the distinction has been tracked down and updated.