Merged PR 84304: Replace ViewManager.onNewTilesReady with TileAdmin events
ViewManager.onNewTilesReady() was used to invalidate the scenes of any on-screen viewports when a tile, its children, or a tile tree was loaded.
Replaced with events on TileAdmin and have ViewManager respond to those events.
I need this so that view attachments can similarly respond (they are not hooked up to ViewManager).