Merged PR 86123: UI: Grow virtualized tree nodes' container width to fit the widest node
Grow virtualized tree nodes' container width to fit the widest node and do not shrink it to avoid horinzontal scrollbar appearing/disappearing.

I intentionally store min width in a ref instead of state to avoid re-rendering the container after we find a wider node. Instead, we store the value in a ref and use it the next time inner container is rendered (after a scroll)

Related work items: #322058